Grids and Datums PE&RS listing

A data.frame of years and months of Grids & Datums column publications by country and country code.

data("GridsDatums")

Format

A data frame with 241 observations on the following 4 variables.

country

name of PE&RS column

month

issue month

year

publication year

ISO

ISO code for country

Details

The journal Photogrammetric Engineering & Remote Sensing, run by the American Society for Photogrammetry and Remote Sensing (ASPRS), began publishing a more-or-less monthly column on the spatial reference systems used in different countries, including their datums. The column first appeared in September 1997, and continued until March 2016; subsequent columns are updated reprints of previous ones. Some also cover other topics, such as world and Martian spatial reference systems. They are written by Clifford J. Mugnier, Louisiana State University, Fellow Emeritus ASPRS. To access the columns, visit https:\/\/www.asprs.org/asprs-publications/grids-and-datums.

Source

https:\/\/www.asprs.org/asprs-publications/grids-and-datums

Examples

data(GridsDatums)
GridsDatums[grep("Norway", GridsDatums$country),]
#>                   country     month year ISO
#> 17  The Kingdom of Norway (October) 1999 NOR
#> 226                Norway    (July) 2017 NOR
GridsDatums[grep("Google", GridsDatums$country),]
#>                             country month year  ISO
#> 201 Datum Origins with Google Earth (May) 2015 <NA>
GridsDatums[grep("^Mars$", GridsDatums$country),]
#>    country     month year  ISO
#> 88    Mars (October) 2005 <NA>