Polygon-class {sp}	R Documentation

Class "Polygon"

Description

class for spatial polygon

Objects from the Class

Objects can be created by calls to the function Polygon

Slots

ringDir:

Object of class "integer"; the ring direction of the ring (polygon) coordinates, holes are expected to be anti-clockwise

labpt:

Object of class "numeric"; an x, y coordinate pair forming the label point of the polygon

area:

Object of class "numeric"; the planar area of the polygon, does not respect projection as objects of this class have no projection defined

hole:

Object of class "logical"; does the polygon seem to be a hole

coords:

Object of class "matrix"; coordinates of the polygon; first point should equal the last point

Extends

Class "Line", directly.

Methods

No methods defined with class "Polygon" in the signature.

Author(s)

Roger Bivand

See Also

Polygons-class, SpatialPolygons-class

[Package sp version 1.5-0 Index]