create objects of class SpatialLines or SpatialLinesDataFrame

create objects of class SpatialLines or SpatialLinesDataFrame from lists of Lines objects and data.frames; extract list od Lines from a SpatialLines object

SpatialLines(LinesList, proj4string = CRS(as.character(NA)))
SpatialLinesDataFrame(sl, data, match.ID = TRUE)
as.SpatialLines.SLDF(SLDF)
getSpatialLinesMidPoints(SL)
LineLength(cc, longlat = FALSE, sum = TRUE)
LinesLength(Ls, longlat = FALSE)
SpatialLinesLengths(SL, longlat)

Arguments

LinesList
list with objects of class Lines-class

proj4string
Object of class "CRS"; holding a valid proj4 string

sl, SL
object of class SpatialLines-class

data
object of class data.frame; the number of rows in data should equal the number of Lines elements in sl

match.ID
logical: (default TRUE): match SpatialLines member Lines ID slot values with data.frame row names, and re-order the data frame rows if necessary; if character: indicates the column in data with Lines IDs to match

SLDF
SpatialLinesDataFrame object

Ls
Object of class Lines

cc
Object of class Line, or two-column matrix with points

longlat
if FALSE, Euclidean distance, if TRUE Great Circle distance in kilometers

sum
logical; if TRUE return scalar length of sum of segments in Line, if FALSE return vector with segment lengths

Value

SpatialLines returns object of class SpatialLines; SpatialLinesDataFrame returns object of class SpatialLinesDataFrame getSpatialLinesMidPoints returns an object of class SpatialPoints,each point containing the (weighted) mean of the lines elements; weighted in the sense that mean is called twice.

See also