SpatialMultiPointsDataFrame-class {sp}	R Documentation

Class "SpatialMultiPointsDataFrame"

Description

Class for spatial attributes that correspond to point sets

Usage

  ## S4 method for signature 'SpatialMultiPointsDataFrame'
x[i, j, ..., drop = TRUE]
  ## S4 method for signature 'SpatialMultiPointsDataFrame,data.frame'
coerce(from, to, strict=TRUE)
  ## S4 method for signature 'SpatialMultiPointsDataFrame'
coordinates(obj)
  ## S4 method for signature 'SpatialMultiPointsDataFrame'
show(object)
  ## S4 method for signature 'SpatialMultiPointsDataFrame'
points(x)

Arguments

x,from,obj,object
SpatialMultiPointsDataFrame object

to
class to which to coerce

strict
see as

i
row indices

j
column indices

drop
see Extract

...
indices passed through

Slots

data:

Object of class data.frame containing the attribute data (may or may not contain the coordinates in its columns)

coords:

Object of class "list"; the list with coordinates matrices; points are rows in the matrix, the list length equals the number of rows in the data slot

bbox:

Object of class "matrix"; bounding box

proj4string:

Object of class "CRS"; projection string

Extends

Class "SpatialMultiPoints", directly. Class "Spatial", by class "SpatialMultiPoints".

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

See Also

coordinates, SpatialMultiPoints-class

Examples

# create three sets of points:
cl1 = cbind(rnorm(3, 10), rnorm(3, 10))
cl2 = cbind(rnorm(5, 10), rnorm(5,  0))
cl3 = cbind(rnorm(7,  0), rnorm(7, 10))

mpdf = SpatialMultiPointsDataFrame(list(cl1, cl2, cl3), data.frame(a = 1:3))
mpdf

plot(mpdf, col = mpdf$a, cex = 1:3)
as(mpdf, "data.frame")
mpdf[1:2,]

[Package sp version 1.5-0 Index]