SpatialPixels-class {sp}	R Documentation

Class "SpatialPixels"

Description

class for defining a pixels, forming a possibly incomplete rectangular grid of arbitrary dimension

Objects from the Class

Objects are created by using e.g.

SpatialPixels(points)

with points of class SpatialPoints-class

Slots

grid

object of class GridTopology-class, defining the grid topology (offset, cellsize, dim)

grid.index

integer; index of points in full grid

coords

coordinates of points, or bbox of grid

bbox:

Object of class "matrix"; bounding box

proj4string:

Object of class "CRS"; projection

Extends

Class "SpatialPoints" directly; Class "Spatial", by class "SpatialPoints".

Methods

coordinates

signature(x = "SpatialPixels"): calculates coordinates for each point on the grid; coordinates are not stored in objects of class SpatialGrid

summary

signature(object = "SpatialPixels"): summarize object

plot

signature(x = "SpatialPixels"): plots cell centers

"["

signature(x = "SpatialPixels"): select pixel cells; the argument drop=FALSE (default) does not recalculate grid topology for the selection, if drop=TRUE the grid topology is recomputed, and might change.

rbind

signature(x = "SpatialPixels"): rbind-like method

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

See Also

SpatialPixelsDataFrame-class, SpatialGrid-class

Examples

data(meuse.grid)
pts = meuse.grid[c("x", "y")]
y = SpatialPixels(SpatialPoints(pts))
class(y)
y
summary(y)
plot(y) # plots grid
plot(y, grid = FALSE) # plots points

[Package sp version 1.5-0 Index]