SpatialPoints {sp}	R Documentation

create objects of class SpatialPoints or SpatialPointsDataFrame

Description

create objects of class SpatialPoints-class or SpatialPointsDataFrame-class from coordinates, and from coordinates and data.frames

Usage

 
SpatialPoints(coords, proj4string=CRS(as.character(NA)), bbox = NULL)
SpatialPointsDataFrame(coords, data, coords.nrs = numeric(0), 
      proj4string = CRS(as.character(NA)), match.ID = TRUE, bbox = NULL)

Arguments

coords
numeric matrix or data.frame with coordinates (each row is a point); in case of SpatialPointsDataFrame an object of class SpatialPoints-class is also allowed

proj4string
projection string of class CRS-class

bbox
bounding box matrix, usually NULL and constructed from the data, but may be passed through for coercion purposes if clearly needed

data
object of class data.frame; the number of rows in data should equal the number of points in the coords object

coords.nrs
numeric; if present, records the column positions where in data the coordinates were taken from (used by coordinates<-)

match.ID
logical; if TRUE AND coords has rownames (i.e., coerced to a matrix, dimnames(coords)[[2]] is not NULL), AND data has row.names (i.e. is a data.frame), then the SpatialPointsDataFrame object is formed by matching the row names of both components, leaving the order of the coordinates in tact. Checks are done to see whether both row names are sufficiently unique, and all data are matched. If FALSE, coordinates and data are simply "glued" together. If character: indicates the column in data with coordinates IDs to match

Value

SpatialPoints returns an object of class SpatialPoints; SpatialPointsDataFrame returns an object of class SpatialPointsDataFrame;

See Also

coordinates, SpatialPoints-class, SpatialPointsDataFrame-class

[Package sp version 1.0-11 Index]