compassRose {sp}	R Documentation

Display a compass rose.

Description

Displays a basic compass rose, usually to orient a map.

Usage

 compassRose(x,y,rot=0,cex=1)

Arguments

x,y
The position of the center of the compass rose in user units.

rot
Rotation for the compass rose in degrees. See Details.

cex
The character expansion to use in the display.

Details

compassRose displays a conventional compass rose at the position requested. The size of the compass rose is determined by the character expansion, as the central "rose" is calculated relative to the character size. Rotation is in degrees counterclockwise.

Value

nil

Author(s)

Jim Lemon

[Package sp version 1.5-0 Index]