coordnames-methods {sp}	R Documentation

retrieve or assign coordinate names for classes in sp

Description

retrieve or assign coordinate names for classes in sp

Methods for coordnames

x = "SpatialPoints"

retrieves coordinate names

x = "SpatialLines"

retrieves coordinate names

x = "Lines"

retrieves coordinate names

x = "Line"

retrieves coordinate names

x = "SpatialPolygons"

retrieves coordinate names

x = "Polygons"

retrieves coordinate names

x = "Polygon"

retrieves coordinate names

Methods for "coordnames<-"

x = "SpatialPoints", value = "character"

assigns coordinate names

x = "SpatialLines", value = "character"

assigns coordinate names

x = "Lines", value = "character"

assigns coordinate names

x = "Line", value = "character"

assigns coordinate names

x = "SpatialPolygons", value = "character"

assigns coordinate names

x = "Polygons", value = "character"

assigns coordinate names

x = "Polygon", value = "character"

assigns coordinate names

x = "GridTopology", value = "character"

assigns coordinate names

x = "SpatialGrid", value = "character"

assigns coordinate names

x = "SpatialPixels", value = "character"

assigns coordinate names

[Package sp version 1.0-11 Index]