disaggregate-methods {sp}	R Documentation

disaggregate SpatialLines, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame objects

Description

disaggregate SpatialLines, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame objects, using functions from rgeos to handle polygon hole nesting

Usage

 disaggregate(x, ...) 

Arguments

x
object of class SpatialLines or SpatialPolygons

...
ignored

Value

object of class SpatialLines or SpatialPolygons, where groups of Line or Polygon are disaggregated to one Line per Lines, or one Polygon per Polygons, respectively.

Author(s)

Robert Hijmans, Edzer Pebesma

[Package sp version 2.1-4 Index]