geometry-methods {sp}	R Documentation

Methods for retrieving the geometry from a composite (geometry + attributes) object

Description

geometry retrieves the SpatialXxx object from a SpatialXxxDataFrame object, with Xxx Lines, Points, Polygons, Grid, or Pixels.

Usage

 geometry(obj) 

Arguments

obj
object of class Spatial

Methods

obj = "Spatial"

obj = "SpatialPointsDataFrame"

obj = "SpatialPolygonsDataFrame"

obj = "SpatialPixelsDataFrame"

obj = "SpatialGridDataFrame"

obj = "SpatialLinesDataFrame"

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

[Package sp version 1.0-11 Index]