is.projected {sp}	R Documentation

Sets or retrieves projection attributes on classes extending SpatialData

Description

Sets or retrieves projection attributes on classes extending SpatialData; set or retrieve option value for error or warning on exceedance of geographical coordinate range, set or retrieve option value for exceedance tolerance of geographical coordinate range. Note that only “+proj=longlat” is accepted for geographical coordinates, which must be ordered (eastings, northings).

Usage

is.projected(obj)
proj4string(obj)
proj4string(obj) <- value
get_ll_warn()
get_ll_TOL()
get_ReplCRS_warn()
set_ll_warn(value)
set_ll_TOL(value)
set_ReplCRS_warn(value)

Arguments

obj
An object of class or extending Spatial-class

value
For proj4string CRS object, containing a valid proj4 string; attempts to assign an object containing “longlat” to data extending beyond longitude [-180, 360] or lattitude [-90, 90] will be stopped. For set_ll_warn a single logical value, if FALSE (default) error on range exceedance, if TRUE, warning. For set_ll_TOL the value of the power of .Machine$double.eps (default 0.25) to use as tolerance in testing range exceedance. set_ReplCRS_warn may be used to turn off warnings issued when changing object CRS with the proj4string replacement method (by setting value=FALSE).

Details

proj4 strings are operative through CRAN package rgdal. For strings defined as “longlat”, the minimum longitude should be -180, the maximum longitude 360, the minimum latitude -90, and the maximum latitude 90. Note that the proj4string replacement method does not project spatial data - for this use spTransform methods in the rgdal package.

Value

is.projected returns a logical; proj4string returns a character vector of length 1; spatial.dimension returns the number of spatial dimensions (2 or 3).

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

See Also

CRS

[Package sp version 1.0-15 Index]