sp-deprecated {sp}	R Documentation

Deprecated functions in sp

Description

Deprecated functions is sp: overlay, getSpP*, getPolygon*, getLines* getSL*

Usage

overlay(x, y, ...)

Arguments

x
first layer

y
second layer, put on top of x

...
optional arguments; see example below

Value

a numerical array of indices of x on locations of y, or a data.frame with (possibly aggregate) properties of x in units of y. (Note that overlay is highly inconsistent, hence its deprecation)

Note

For overlay the new implementation is found in the over method, and works somewhat different (because consistent).

[Package sp version 1.0-11 Index]