spTransform {sp}	R Documentation

spTransform for map projection and datum transformation

Description

spTransform for map projection and datum transformation

Usage

spTransform(x, CRSobj, ...)

Arguments

x
object to be transformed

CRSobj
object of class CRS

...
further arguments

Value

object with coordinates transformed to the new coordinate reference system.

Note

Package sp only provides the definition of the generic, methods doing actual transformation are found in package rgdal, see spTransform

[Package sp version 1.0-15 Index]