Polygons-class {sp}	R Documentation

Class "Polygons"

Description

Collection of objects of class "Polygon"

Objects from the Class

Objects can be created by calls to the function Polygons

Slots

Polygons:

Object of class "list"; list with objects of class Polygon-class

plotOrder:

Object of class "integer"; order in which the Polygon objects should be plotted, currently by order of decreasing size

labpt:

Object of class "numeric"; pair of x, y coordinates giving a label point, the label point of the largest polygon component

ID:

Object of class "character"; unique identifier string

area:

Object of class "numeric"; the gross total planar area of the Polygon list but not double-counting holes (changed from 0.9-58 - islands are summed, holes are ignored rather than subtracted); these values are used to make sure that polygons of a smaller area are plotted after polygons of a larger area, does not respect projection as objects of this class have no projection defined

Methods

No methods defined with class "Polygons" in the signature.

Note

By default, single polygons (where Polygons is a list of length one) are not expected to be holes, but in multiple polygons, hole definitions for member polygons can be set. Polygon objects belonging to an Polygons object should either not overlap one-other, or should be fully included (as lakes or islands in lakes). They should not be self-intersecting. Checking of hole FALSE/TRUE status for Polygons objects is included in the maptools package using functions in the rgeos package, function checkPolygonsHoles().

Author(s)

Roger Bivand

[Package sp version 1.5-0 Index]