SpatialGrid-class {sp}	R Documentation

Class "SpatialGrid"

Description

class for defining a full, rectangular grid of arbitrary dimension

Objects from the Class

Objects are created by using e.g.

SpatialGrid(grid)

with grid of class GridTopology-class

Slots

grid

object of class GridTopology-class, defining the grid topology (offset, cellsize, dim)

bbox:

Object of class "matrix"; bounding box

proj4string:

Object of class "CRS"; projection

Extends

Class "SpatialPoints" directly; Class "Spatial", by class "SpatialPoints".

Methods

coordinates

signature(x = "SpatialGrid"): calculates coordinates for each point on the grid; coordinates are not stored in objects of class SpatialGrid

summary

signature(object = "SpatialGrid"): summarize object

plot

signature(x = "SpatialGrid"): plots cell centers

"["

signature(x = "SpatialGrid"): select rows and columns

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

See Also

SpatialGridDataFrame-class, SpatialGrid

Examples

x = GridTopology(c(0,0), c(1,1), c(5,5))
class(x)
x
summary(x)
coordinates(x)
y = SpatialGrid(grid = x)
class(y)
y

[Package sp version 1.5-0 Index]