a class for spatial lines with attributes

this class holds data consisting of (sets of lines), where each set of lines relates to an attribute row in a data.frame

Objects from the Class

can be created by the function SpatialLinesDataFrame

Slots

data:

Object of class data.frame containing the attribute table

lines:

Object of class "list"; see SpatialLines-class

bbox:

Object of class "matrix"; see Spatial-class

proj4string:

Object of class "CRS"; see CRS-class

Extends

Class "SpatialLines", directly. Class "Spatial", by class "SpatialLines".

Methods

Methods defined with class "SpatialLinesDataFrame" in the signature:

[

signature(x = "SpatialLinesDataFrame"): subset rows or columns; in case of row subsetting, the line sets are also subsetted; NAs are not permitted in the row index

coordinates

signature(obj = "SpatialLinesDataFrame"): retrieves a list with lists of coordinate matrices

show

signature(object = "SpatialLinesDataFrame"): print method

plot

signature(x = "SpatialLinesDataFrame"): plot points

lines

signature(object = "SpatialLinesDataFrame"): add lines to plot

rbind

signature(object = "SpatialLinesDataFrame"): rbind-like method

Note

rbind for SpatialLinesDataFrame is only possible for objects with unique IDs. If you want to rbind objects with duplicated IDs, seespChFIDs.

Author

Roger Bivand; Edzer Pebesma

See also