geometry-methods {sp}	R Documentation

Methods for retrieving the geometry from a composite (geometry + attributes) object

Description

geometry retrieves the SpatialXxx object from a SpatialXxxDataFrame object, with Xxx Lines, Points, Polygons, Grid, or Pixels. geometry<- converts a data.frame into a Spatial object.

Usage

 
geometry(obj) 
geometry(obj) <- value 

Arguments

obj
in case of assignment, a data.frame, else an object of class Spatial

value
object of class Spatial

Methods

obj = "Spatial"

obj = "SpatialPointsDataFrame"

obj = "SpatialMultiPointsDataFrame"

obj = "SpatialPolygonsDataFrame"

obj = "SpatialPixelsDataFrame"

obj = "SpatialGridDataFrame"

obj = "SpatialLinesDataFrame"

Author(s)

Edzer Pebesma, edzer.pebesma@uni-muenster.de

Examples

data(meuse)
m = meuse
coordinates(m) = meuse[, c("x", "y")]
pts = geometry(m)
class(pts)
geometry(meuse) = pts
class(meuse)
identical(m, meuse) # TRUE

[Package sp version 1.5-0 Index]