Meuse river data set

This data set gives locations and topsoil heavy metal concentrations, along with a number of soil and landscape variables at the observation locations, collected in a flood plain of the river Meuse, near the village of Stein (NL). Heavy metal concentrations are from composite samples of an area of approximately 15 m x 15 m.

data(meuse)

Format

This data frame contains the following columns:

x

a numeric vector; Easting (m) in Rijksdriehoek (RDH) (Netherlands topographical) map coordinates

y

a numeric vector; Northing (m) in RDH coordinates

cadmium

topsoil cadmium concentration, mg kg-1 soil ("ppm"); zero cadmium values in the original data set have been shifted to 0.2 (half the lowest non-zero value)

copper

topsoil copper concentration, mg kg-1 soil ("ppm")

lead

topsoil lead concentration, mg kg-1 soil ("ppm")

zinc

topsoil zinc concentration, mg kg-1 soil ("ppm")

elev

relative elevation above local river bed, m

dist

distance to the Meuse; obtained from the nearest cell in meuse.grid, which in turn was derived by a spread (spatial distance) GIS operation, horizontal precision 20 metres; then normalized to $[0,1]$

om

organic matter, kg (100 kg)-1 soil (percent)

ffreq

flooding frequency class: 1 = once in two years; 2 = once in ten years; 3 = one in 50 years

soil

soil type according to the 1:50 000 soil map of the Netherlands. 1 = Rd10A (Calcareous weakly-developed meadow soils, light sandy clay); 2 = Rd90C/VII (Non-calcareous weakly-developed meadow soils, heavy sandy clay to light clay); 3 = Bkd26/VII (Red Brick soil, fine-sandy, silty light clay)

lime

lime class: 0 = absent, 1 = present by field test with 5% HCl

landuse

landuse class: Aa Agriculture/unspecified = , Ab = Agr/sugar beetsm, Ag = Agr/small grains, Ah = Agr/??, Am = Agr/maize, B = woods, Bw = trees in pasture, DEN = ??, Fh = tall fruit trees, Fl = low fruit trees; Fw = fruit trees in pasture, Ga = home gardens, SPO = sport field, STA = stable yard, Tv = ?? , W = pasture

dist.m

distance to river Meuse in metres, as obtained during the field survey

Note

row.names refer to the original sample number.

Soil units were mapped with a minimum delination width of 150 m, and so somewhat generalize the landscape.

Approximate equivalent World Reference Base 2002 for Soil Resources names are: Rd10A Gleyic Fluvisols; Rd90C Haplic Fluvisols; Bkd26 Haplic Luvisols. Units Rd90C and Bkd26 have winter groundwater > 80cm, summer > 120cm depth.

Author

Field data were collected by Ruud van Rijn and Mathieu Rikken; compiled for R by Edzer Pebesma; description extended by David Rossiter

References

M G J Rikken and R P G Van Rijn, 1993. Soil pollution with heavy metals - an inquiry into spatial variation, cost of mapping and the risk evaluation of copper, cadmium, lead and zinc in the floodplains of the Meuse west of Stein, the Netherlands. Doctoraalveldwerkverslag, Dept. of Physical Geography, Utrecht University

P.A. Burrough, R.A. McDonnell, 1998. Principles of Geographical Information Systems. Oxford University Press.

Stichting voor Bodemkartering (STIBOKA), 1970. Bodemkaart van Nederland : Blad 59 Peer, Blad 60 West en 60 Oost Sittard: schaal 1 : 50 000. Wageningen, STIBOKA.

http://www.gstat.org/

Examples

data(meuse)
summary(meuse)
#>        x                y             cadmium           copper      
#>  Min.   :178605   Min.   :329714   Min.   : 0.200   Min.   : 14.00  
#>  1st Qu.:179371   1st Qu.:330762   1st Qu.: 0.800   1st Qu.: 23.00  
#>  Median :179991   Median :331633   Median : 2.100   Median : 31.00  
#>  Mean   :180005   Mean   :331635   Mean   : 3.246   Mean   : 40.32  
#>  3rd Qu.:180630   3rd Qu.:332463   3rd Qu.: 3.850   3rd Qu.: 49.50  
#>  Max.   :181390   Max.   :333611   Max.   :18.100   Max.   :128.00  
#>                                                                     
#>       lead            zinc             elev             dist        
#>  Min.   : 37.0   Min.   : 113.0   Min.   : 5.180   Min.   :0.00000  
#>  1st Qu.: 72.5   1st Qu.: 198.0   1st Qu.: 7.546   1st Qu.:0.07569  
#>  Median :123.0   Median : 326.0   Median : 8.180   Median :0.21184  
#>  Mean   :153.4   Mean   : 469.7   Mean   : 8.165   Mean   :0.24002  
#>  3rd Qu.:207.0   3rd Qu.: 674.5   3rd Qu.: 8.955   3rd Qu.:0.36407  
#>  Max.   :654.0   Max.   :1839.0   Max.   :10.520   Max.   :0.88039  
#>                                                                     
#>        om         ffreq  soil   lime       landuse       dist.m      
#>  Min.   : 1.000   1:84   1:97   0:111   W      :50   Min.   :  10.0  
#>  1st Qu.: 5.300   2:48   2:46   1: 44   Ah     :39   1st Qu.:  80.0  
#>  Median : 6.900   3:23   3:12           Am     :22   Median : 270.0  
#>  Mean   : 7.478                         Fw     :10   Mean   : 290.3  
#>  3rd Qu.: 9.000                         Ab     : 8   3rd Qu.: 450.0  
#>  Max.   :17.000                         (Other):25   Max.   :1000.0  
#>  NA's   :2                              NA's   : 1                   
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")