spTransform {sp}	R Documentation

spTransform for map projection and datum transformation

Description

spTransform for map projection and datum transformation

Usage

spTransform(x, CRSobj, ...)

Arguments

x
object to be transformed

CRSobj
object of class CRS, or of class character in which case it is converted to CRS

...
further arguments (ignored)

Value

object with coordinates transformed to the new coordinate reference system.

Note

Package rgdal provides the methods doing actual transformation, see spTransform; when rgdal cannot be loaded, an error message follows.

[Package sp version 1.5-0 Index]